+++ src/create_sb.py
@@ -7,7 +7,7 @@ import db
 __author__ = 'jmrodriguezc'
 __credits__ = ["Jose Rodriguez", "Jesus Vazquez"]
 __license__ = "Creative Commons Attribution-NonCommercial-NoDerivs 4.0 Unported License https://creativecommons.org/licenses/by-nc-nd/4.0/"
-__version__ = "2.2"
+__version__ = "2.4"
 __maintainer__ = "Jose Rodriguez"
 __email__ = "jmrodriguezc@cnic.es"
 __status__ = "Development"