+++ changelog.md
@@ -1,4 +1,17 @@
 ___
+## v2.4
+```
+DATE: 2021_05
+```
+
++ Fixing some bugs in the parser section.
+
++ OMIM database has been included.
+
++ The following filter in the GO terms is by default:
+"cat_GO_*:EXP,IDA,IPI,IMP,IGI,IEP,HTP,HDA,HMP,HGI,HEP,IBA,IBD,IKR,IRD"
+
+___
 ## v2.3
 ```
 DATE: 2021_05